code to run matching with consumer thread -> step detection will catch up if you wait for the calculations to finish as steps are buffered. steps are now virtual, i.e. we work with varying step lengths whichare then buffered and then added to the consumer queue compass needle removed, compass shows north, android looks into the direction the user looks (might be more intuitive)

+++ src/de/uvwxy/footpath2/matching/multifit/PartialPenaltyTree.java
@@ -70,8 +70,8 @@ public class PartialPenaltyTree implements DrawToCanvas {
 		return this.virtualStepLength;
 	}
 
-	public void onStepUpdate(float bearing, double steplength, long timestamp, double estimatedStepLengthError,
-			double estimatedBearingError) {
+	public void onStepUpdate(float bearing, float steplength, long timestamp, float estimatedStepLengthError,
+			float estimatedBearingError) {
 		currentStep++;
 		bearings.add(new Float(bearing));
 		int deletedNodes = 0;
