+++ src/de/uvwxy/footpath2/movement/StepEventListener.java
@@ -1,6 +1,6 @@
 package de.uvwxy.footpath2.movement;
 
 public interface StepEventListener {
-	void onStepUpdate(double bearing, double steplength, long timestamp,
-			double estimatedStepLengthError, double estimatedBearingError);
+	void onStepUpdate(float bearing, float steplength, long timestamp,
+			float estimatedStepLengthError, float estimatedBearingError);
 }
